requirements.txt.txt
torch
torchvision
tensorboardX
tensorflow-gpu